moving out npy_to_blocks_list to base_util
